+++ commons/constants/index.ts
@@ -21,10 +21,13 @@ export type { RouteKey, RoutePath } from './routes';
 export { API_ENDPOINTS } from './endpoints';
 export type { ApiEndpointKey, ApiEndpointPath } from './endpoints';
 
+// 미디어 업로드 상수
+export { ALLOWED_EXTENSIONS, MIME_TYPE_MAP, SIZE_LIMITS } from './media';
+
 // 디자인 토큰 (tailwind.config.js와 동기화)
 export { BorderRadius } from './borderRadius';
 export { Colors } from './color';
 export { FontFamily, FontSize, FontWeight, LineHeight } from './fonts';
 export { Spacing } from './spacing';
-export { getTypographyStyle, Typography } from './typography';
+export { Typography, getTypographyStyle } from './typography';
 export type { TypographyKey } from './typography';
