+++ commons/hooks/index.ts
@@ -3,4 +3,5 @@
  * 공통 커스텀 훅 export
  */
 
+export { useMediaUpload } from './useMediaUpload';
 export { useNavigation } from './useNavigation';
